refined data type

+++ ESP32/Software/Configuration/Blynk_example.h
@@ -1,2 +1,2 @@
-const int BLYNK_PORT = 8080;
+const unsigned short int BLYNK_PORT = 8080;
 const bool BLYNK_USE_LOCAL_SERVER = false;
